Move common component to intellij plugin lib
