+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-plugin-lib/build.gradle
@@ -16,6 +16,7 @@ sourceSets {
 }
 
 dependencies {
+    compile 'com.microsoft.azure:azure-toolkit-resource-lib:0.12.0-SNAPSHOT'
     compile 'com.microsoft.azure:azure-toolkit-auth-lib:0.12.0-SNAPSHOT'
     compile 'com.microsoft.azure:azure-toolkit-ide-common-lib:0.12.0-SNAPSHOT'
     compile 'org.dom4j:dom4j:2.1.3', {
