+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-plugin-lib/src/main/java/com/microsoft/azure/toolkit/intellij/common/component/resourcegroup/ResourceGroupCreationDialog.form
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="com.microsoft.azure.toolkit.intellij.appservice.resourcegroup.ResourceGroupCreationDialog">
+<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="com.microsoft.azure.toolkit.intellij.common.component.resourcegroup.ResourceGroupCreationDialog">
   <grid id="27dc6" binding="contentPanel" layout-manager="GridLayoutManager" row-count="3" column-count="2" same-size-horizontally="false" same-size-vertically="false" hgap="-1" vgap="-1">
     <margin top="0" left="0" bottom="0" right="0"/>
     <constraints>
@@ -19,7 +19,7 @@
           <text value="Name:"/>
         </properties>
       </component>
-      <component id="c8aa9" class="com.microsoft.azure.toolkit.intellij.appservice.resourcegroup.ResourceGroupNameTextField" binding="textName">
+      <component id="c8aa9" class="com.microsoft.azure.toolkit.intellij.common.component.resourcegroup.ResourceGroupNameTextField" binding="textName">
         <constraints>
           <grid row="1" column="1" row-span="1" col-span="1" vsize-policy="3" hsize-policy="3" anchor="0" fill="1" indent="0" use-parent-layout="false"/>
         </constraints>
