+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-plugin-lib/src/main/java/com/microsoft/azure/toolkit/intellij/common/component/resourcegroup/ResourceGroupNameTextField.java
@@ -2,7 +2,7 @@
  * Copyright (c) Microsoft Corporation. All rights reserved.
  * Licensed under the MIT License. See License.txt in the project root for license information.
  */
-package com.microsoft.azure.toolkit.intellij.appservice.resourcegroup;
+package com.microsoft.azure.toolkit.intellij.common.component.resourcegroup;
 
 import com.microsoft.azure.CloudException;
 import com.microsoft.azure.toolkit.intellij.common.ValidationDebouncedTextInput;
