+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-plugin-lib/src/main/resources/com/microsoft/azure/toolkit/message/common.properties
@@ -1,0 +1,3 @@
+common.resourceGroup.description=A resource group is a container that holds related resources for an Azure solution.
+common.resourceGroup.create.title=New Resource Group
+common.resourceGroup.create.tooltip=Create new resource group
