+++ PluginsAndFeatures/azure-toolkit-for-intellij/src/main/java/com/microsoft/azure/toolkit/intellij/appservice/AppServiceInfoAdvancedPanel.form
@@ -157,7 +157,7 @@
           <text value="Artifact(Optional):"/>
         </properties>
       </component>
-      <component id="8161" class="com.microsoft.azure.toolkit.intellij.appservice.resourcegroup.ResourceGroupComboBox" binding="selectorGroup">
+      <component id="8161" class="com.microsoft.azure.toolkit.intellij.common.component.resourcegroup.ResourceGroupComboBox" binding="selectorGroup">
         <constraints>
           <grid row="2" column="1" row-span="1" col-span="2" vsize-policy="3" hsize-policy="3" anchor="0" fill="1" indent="0" use-parent-layout="false"/>
         </constraints>
