+++ PluginsAndFeatures/azure-toolkit-for-intellij/src/main/java/com/microsoft/azure/toolkit/intellij/mysql/creation/MySQLCreationAdvancedPanel.form
@@ -43,7 +43,7 @@
           <text value="Resource group:"/>
         </properties>
       </component>
-      <component id="ddd13" class="com.microsoft.azure.toolkit.intellij.appservice.resourcegroup.ResourceGroupComboBox" binding="resourceGroupComboBox">
+      <component id="ddd13" class="com.microsoft.azure.toolkit.intellij.common.component.resourcegroup.ResourceGroupComboBox" binding="resourceGroupComboBox">
         <constraints>
           <grid row="2" column="1" row-span="1" col-span="1" vsize-policy="0" hsize-policy="2" anchor="8" fill="1" indent="1" use-parent-layout="false"/>
         </constraints>
